refactor: Simplify exposed layout
